+++ cmd/reliably/slo/report/cmd.go
@@ -39,6 +39,7 @@ type ReportOptions struct {
 	ManifestPath  string
 	OutputPath    string
 	OutputFormat  string
+	TemplateFile  string
 	WatchFlag     bool
 	OutputPaths   []string
 	OutputFormats []string
@@ -50,7 +51,7 @@ type ReportOptions struct {
 const defaultFormat = "table"
 
 var (
-	supportedFormats  = Choice{"json", "yaml", "text", "table", "markdown"}
+	supportedFormats  = Choice{"json", "yaml", "text", "table", "markdown", "template"}
 	deprecatedFormats = Choice{"simple", "tabbed"}
 )
 
@@ -69,10 +70,22 @@ formats at once, with using '--format' and '--output' flags with
 comma-separated list as values.`),
 		Example: `  $ reliably slo report
   $ reliably slo report -f text
+  $ reliably slo report -t slo-report.tmpl
+  $ reliably slo report -t slo-report.tmpl -o slo-report.txt
   $ reliably slo report -f markdown -o report.md
   $ reliably slo report -f yaml,json -o o.yaml,o.json`,
 		PreRunE: func(cmd *cobra.Command, args []string) error {
 			// Validate command options
+
+			if opts.TemplateFile != "" {
+				// check template file exists, see outputFormat to template
+				_, err := os.Open(opts.TemplateFile) // For read access.
+				if err != nil {
+					return fmt.Errorf("Error opening template file: %s", err)
+				}
+
+				opts.OutputFormat = "template"
+			}
 
 			if opts.OutputFormat != "" {
 				opts.OutputFormats = strings.Split(opts.OutputFormat, ",")
@@ -125,6 +138,8 @@ comma-separated list as values.`),
 					format = report.SimpleText
 				case "markdown":
 					format = report.MARKDOWN
+				case "template":
+					format = report.TEMPLATE
 				case "yaml":
 					format = report.YAML
 				}
@@ -152,6 +167,7 @@ comma-separated list as values.`),
 
 	cmd.Flags().StringVarP(&opts.ManifestPath, "manifest", "m", manifest.DefaultManifestPath, "the location of the manifest file")
 	cmd.Flags().StringVarP(&opts.OutputPath, "output", "o", "", "where the report should be written to")
+	cmd.Flags().StringVarP(&opts.TemplateFile, "template", "t", "", "the name of the template to use for the report output")
 	cmd.Flags().StringVarP(&opts.OutputFormat, "format", "f", "table", fmt.Sprintf("specify the report format. Allowed Values: %v", supportedFormats))
 	cmd.Flags().BoolVarP(&opts.WatchFlag, "watch", "w", false, "continuously watch for changes in report output")
 	cmd.Flags().StringVar(&opts.Service, "service", "", "the name of the service")
@@ -224,7 +240,7 @@ func reportRun(opts *ReportOptions) error {
 			w = outfile
 			// we cannot defer outfile closing here as we are in a for-loop
 		}
-		report.Write(out.Format, r, w, log.StandardLogger(), &lr, &reports)
+		report.Write(out.Format, r, w, opts.TemplateFile, log.StandardLogger(), &lr, &reports)
 
 		if outfile, ok := w.(*os.File); ok {
 			outfile.Close() // explicitly closing the file handle
@@ -284,7 +300,7 @@ func watch(opts *ReportOptions) error {
 		case r := <-rChan:
 			clearScreen()
 			fmt.Println(color.Magenta("Refreshing SLO report every 3 seconds."), "Press CTRL+C to quit.")
-			report.Write(report.TABBED, r, os.Stdout, log.StandardLogger(), last, nil)
+			report.Write(report.TABBED, r, os.Stdout, "", log.StandardLogger(), last, nil)
 			last = r
 
 		case err := <-errChan: